Добавлена система сохранения текущих номеров вкладок. Удалена старая система сохранения значений. Сохранение настроек теперь выполняется с отступом в 4 пробела. Добавлено автоматическое подключение из вкладки с командами.

+++ src/Windows/Controllers/TestControllers/UnitTestsController.cpp
@@ -641,7 +641,9 @@ void UnitTestsController::redrawStatistics()
         else
         {
             // Add table element
-            element.second.second = new QTableWidgetItem(element.second.first);
+            element.second.second = new QTableWidgetItem(
+                    (int) element.second.first
+            );
 
             ui()->unitTestsStatisticsTableWidget->setRowCount(
                     ui()->unitTestsStatisticsTableWidget->rowCount() + 1
